requirements.txt
# 核心依赖
# Python >= 3.9 required

# 数据库
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.0
sqlalchemy==2.0.25
alembic==1.13.1

# 数据获取
tushare==1.2.89
akshare==1.12.75
baostock==0.8.80
pandas==2.2.1
numpy==1.26.4

# 数据分析
ta-lib==0.4.28
scikit-learn==1.4.1
lightgbm==4.3.0
statsmodels==0.14.1

# 技术分析库
pandas-ta==0.3.14
technical==0.0.1

# Web框架
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# 任务调度
celery==5.3.6
redis==5.0.1

# 消息队列
pika==1.3.2

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 监控
prometheus-client==0.19.0

# 可视化
plotly==5.19.0
matplotlib==3.8.3

# 开发工具
black==24.1.1
isort==5.13.2
pytest==8.0.0
pytest-asyncio==0.23.5

# OpenClaw集成
openclaw==2026.2.13